work_step4_ozonfix_v6/requirements.txt
PyQt6>=6.5
openpyxl>=3.1
python-barcode[images]>=0.15
reportlab>=4.0
Pillow>=10.0
pyinstaller>=6.0
